# backend/app/swagger.yaml
openapi: 3.0.0
info:
  title: Kinri API
  version: 1.0.0
  description: |
    This is the Kinri API documentation for external partners. All endpoints are prefixed with `/api` and some require authentication via Bearer JWT (see security section).
servers:
  - url: /api
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /assessments/evaluate:
    post:
      tags: [Assessments]
      summary: Evaluate assessment responses and get AI feedback
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
      responses:
        '200':
          description: Assessment evaluation result
        '400':
          description: Bad request
        '401':
          description: Unauthorized
  /assessments/assessment_questions:
    get:
      tags: [Assessments]
      summary: Get all assessment questions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of assessment questions
        '401':
          description: Unauthorized
  /assessments/condition_weights:
    get:
      tags: [Assessments]
      summary: Get question-condition weights
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Grouped question-condition weights
        '401':
          description: Unauthorized
...